chore(entities): refactor all imports to match new path mappings
