+++ src/entities/Donor.ts
@@ -1,6 +1,6 @@
 import { Entity } from 'typeorm'
 
-import { User } from '@entities'
+import { User } from '@entities/User'
 
 @Entity('donors')
 export class Donor extends User {}
