+++ src/entities/Ong.ts
@@ -1,6 +1,6 @@
 import { Entity } from 'typeorm'
 
-import { User } from '@entities'
+import { User } from '@entities/User'
 
 @Entity('ongs')
 export class Ong extends User {}
