--- src/entities/index.ts
@@ -1,6 +1,0 @@
-export * from './BaseEntity'
-export * from './Incident'
-export * from './User'
-export * from './Donor'
-export * from './Ong'
-export * from './Donation'
